prov/efa: move and rename efa_rdm_error_write_msg()
Move the function efa_rdm_error_write_msg() to efa_rdm_util.c
and rename it efa_rdm_write_error_msg().
