+++ prov/efa/Makefile.include
@@ -109,7 +109,6 @@ _efa_headers = \
 	prov/efa/src/rdm/efa_rdm_protocol.h \
 	prov/efa/src/rdm/rxr_tp_def.h \
 	prov/efa/src/rdm/rxr_tp.h \
-	prov/efa/src/rdm/efa_rdm_error.h \
 	prov/efa/src/rdm/efa_rdm_srx.h \
 	prov/efa/src/rdm/efa_rdm_util.h
